+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- `RichChat` component for local chats with embeddable content
+### Changed
+- chat avatars align with the top of multi-line messages
 
 ## [0.18.1]
 - Added `Markdown` component